added awsconfig method in utils
