+++ aws-kinesis-http-connector/go.mod
@@ -1,0 +1,10 @@
+module github.com/fission/keda-connectors/aws-kinesis-http-connector
+
+go 1.13
+
+require (
+	github.com/aws/aws-sdk-go v1.34.24
+	github.com/fission/keda-connectors/common v0.0.0-20200915102844-c68eb4e4d582
+	github.com/joho/godotenv v1.3.0
+	go.uber.org/zap v1.16.0
+)
