+++ aws-kinesis-http-connector/main.go
@@ -11,10 +11,11 @@ import (
 	"sync"
 	"time"
 
+	"github.com/fission/keda-connectors/awsutil"
+
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/kinesis"
-	"github.com/fission/keda-connectors/awsutil"
 	"github.com/fission/keda-connectors/common"
 
 	"github.com/joho/godotenv"
