--- awsutil/go.mod
@@ -1,8 +1,0 @@
-module github.com/fission/keda-connectors/awsutil
-
-go 1.13
-
-require (
-	github.com/aws/aws-sdk-go v1.34.24
-	github.com/pkg/errors v0.9.1
-)
