+++ common/go.mod
@@ -3,6 +3,7 @@ module github.com/fission/keda-connectors/common
 go 1.12
 
 require (
+	github.com/aws/aws-sdk-go v1.34.24
 	github.com/pkg/errors v0.9.1
 	go.uber.org/zap v1.15.0
 )
